Ahora MapParser compila. Ahora el Editor agrega cajas y zonas de bomba de tamaño 64x64

+++ Common/YamlParser/MapConfigParser.h
@@ -1,10 +1,10 @@
 #ifndef MAP_PARSER_H
 #define MAP_PARSER_H
-/*
+
 #include <string>
 #include <iostream>
 #include "yaml-cpp/yaml.h"
-#include "Position.h"
+#include "../../Server/src/GameWorld/WorldObjects/GroundMap/Position.h"
 #include "WeaponType.h"
 
 class MapParser {
@@ -13,7 +13,7 @@ class MapParser {
         int pos_x;
         int pos_y;
         std::string type;
-        int it;
+        size_t it;
 
     public:
         MapParser(const std::string &file_name);
@@ -31,5 +31,4 @@ class MapParser {
         ~MapParser();
 };
 
-*/
 #endif
